修复订单小数部分丢失的bug，修复支付后显示bug

+++ common/src/main/java/com/minipay/common/req/OrderReq.java
@@ -1,8 +1,10 @@
 package com.minipay.common.req;
 
+import java.math.BigDecimal;
+
 public class OrderReq {
     private Long orderId;
-    private Long amount;
+    private BigDecimal amount;
 
     public Long getOrderId() {
         return orderId;
@@ -12,11 +14,11 @@ public class OrderReq {
         this.orderId = orderId;
     }
 
-    public Long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
